dags/easy_com/requirements.txt
pandas==2.2.3
pyarrow==18.0.0
pybigquery==0.10.2
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.41.2
google-auth==2.35.0
